requirements.txt
# ── AI providers ───────────────────────────────────────────────────────────────
anthropic==0.97.0
openai==2.32.0

# ── Web UI ─────────────────────────────────────────────────────────────────────
streamlit==1.57.0

# ── Data ───────────────────────────────────────────────────────────────────────
yfinance==1.3.0
pandas==3.0.2
numpy==2.4.4
requests==2.33.1

# ── HTML/XML parsing ───────────────────────────────────────────────────────────
lxml==6.1.0
beautifulsoup4==4.14.3

# ── PDF generation ─────────────────────────────────────────────────────────────
reportlab==4.5.0

# ── Config ─────────────────────────────────────────────────────────────────────
python-dotenv==1.2.2
